Rename variables to vpc_remote_state_* for consistency
Also, remove the comment asking whether a single DynamoDB table can
be used for locking of remote state. The answer is yes, a single
table works just fine because each state file has a different key
in the S3 bucket when storing remote state and that same key is used
as the ID in the DynamoDB table for locking.